moves shortcut helpers to core-utils + uses it everywhere
