+++ webapp/packages/core-utils/src/IKeyBinding.ts
@@ -1,0 +1,15 @@
+/*
+ * CloudBeaver - Cloud Database Manager
+ * Copyright (C) 2020-2025 DBeaver Corp and others
+ *
+ * Licensed under the Apache License, Version 2.0.
+ * you may not use this file except in compliance with the License.
+ */
+
+export interface IKeyBinding {
+  id: string;
+  preventDefault?: boolean;
+  keys?: string | string[];
+  keysWin?: string | string[];
+  keysMac?: string | string[];
+}
