+++ webapp/packages/core-utils/src/formatKeyToDisplayKey.ts
@@ -1,0 +1,53 @@
+/*
+ * CloudBeaver - Cloud Database Manager
+ * Copyright (C) 2020-2025 DBeaver Corp and others
+ *
+ * Licensed under the Apache License, Version 2.0.
+ * you may not use this file except in compliance with the License.
+ */
+import { getOS, OperatingSystem } from './getOS.js';
+
+const FORMAT_SHORTCUT_KEYS_MAP: Record<string, string> = {
+  comma: ',',
+  slash: '/',
+  backslash: '\\',
+  backspace: '⌫',
+  tab: 'tab',
+  clear: 'clear',
+  enter: '↵',
+  return: '↵',
+  escape: 'escape',
+  esc: 'escape',
+  space: '␣',
+  up: '↑',
+  down: '↓',
+  left: '←',
+  right: '→',
+  pageup: 'pageup',
+  pagedown: 'pagedown',
+  del: '⌦',
+  delete: '⌦',
+};
+
+export function formatKeyToDisplayKey(code: string): string {
+  const lowerCaseCode = code.toLowerCase();
+  const OS = getOS();
+
+  switch (lowerCaseCode) {
+    case 'mod':
+      if (OS === OperatingSystem.windowsOS || OS === OperatingSystem.linuxOS) {
+        return 'CTRL';
+      }
+      if (OS === OperatingSystem.macOS) {
+        return 'CMD';
+      }
+      return code;
+    case 'alt':
+      if (OS === OperatingSystem.macOS) {
+        return 'OPTION';
+      }
+      return 'ALT';
+    default:
+      return FORMAT_SHORTCUT_KEYS_MAP[lowerCaseCode] ?? code;
+  }
+}
