+++ webapp/packages/core-utils/src/index.ts
@@ -87,3 +87,9 @@ export * from './bindFunctions.js';
 export * from './getDomainFromUrl.js';
 export * from './isNumber.js';
 export * from './getSubjectDifferences.js';
+export * from './formatKeyToDisplayKey.js';
+export * from './transformKeys.js';
+export * from './getCommonAndOSSpecificKeys.js';
+export * from './getOSSpecificKeys.js';
+export * from './getShortcutStringFromBinding.js';
+export * from './IKeyBinding.js';
