+++ webapp/packages/core-view/src/index.ts
@@ -44,8 +44,6 @@ export * from './Action/IActionHandler.js';
 export * from './Action/IActionInfo.js';
 export * from './Action/IActionItem.js';
 export * from './Action/KeyBinding/IKeyBinding.js';
-export * from './Action/KeyBinding/getCommonAndOSSpecificKeys.js';
-export * from './Action/KeyBinding/getBindingLabel.js';
 export * from './LoadableStateContext/DATA_CONTEXT_LOADABLE_STATE.js';
 export * from './Menu/MenuItem/IMenuCheckboxItem.js';
 export * from './Menu/MenuItem/IMenuActionItem.js';
