+++ webapp/packages/plugin-object-viewer-nav-tree-link/src/ObjectViewerNavTreeLinkMenuService.ts
@@ -1,6 +1,6 @@
 /*
  * CloudBeaver - Cloud Database Manager
- * Copyright (C) 2020-2024 DBeaver Corp and others
+ * Copyright (C) 2020-2025 DBeaver Corp and others
  *
  * Licensed under the Apache License, Version 2.0.
  * you may not use this file except in compliance with the License.
@@ -9,11 +9,12 @@ import type { IDataContextProvider } from '@cloudbeaver/core-data-context';
 import { NavigationTreeService, DATA_CONTEXT_ELEMENTS_TREE, MENU_ELEMENTS_TREE_TOOLS } from '@cloudbeaver/plugin-navigation-tree';
 import { Bootstrap, injectable } from '@cloudbeaver/core-di';
 import { LocalizationService } from '@cloudbeaver/core-localization';
-import { ActionService, getBindingLabel, type IAction, KeyBindingService, MenuService } from '@cloudbeaver/core-view';
+import { ActionService, type IAction, KeyBindingService, MenuService } from '@cloudbeaver/core-view';
 import { ConnectionSchemaManagerService } from '@cloudbeaver/plugin-datasource-context-switch';
 
 import { ACTION_LINK_OBJECT } from './ACTION_LINK_OBJECT.js';
 import { KEY_BINDING_LINK_OBJECT } from './KEY_BINDING_LINK_OBJECT.js';
+import { getShortcutStringFromBinding } from '@cloudbeaver/core-utils';
 
 @injectable()
 export class ObjectViewerNavTreeLinkMenuService extends Bootstrap {
@@ -38,7 +39,7 @@ export class ObjectViewerNavTreeLinkMenuService extends Bootstrap {
       getActionInfo: (context, action) => {
         switch (action) {
           case ACTION_LINK_OBJECT: {
-            const bindingLabel = getBindingLabel(KEY_BINDING_LINK_OBJECT);
+            const bindingLabel = getShortcutStringFromBinding(KEY_BINDING_LINK_OBJECT);
             const tooltip =
               this.localizationService.translate('app_navigationTree_action_link_with_editor') + (bindingLabel ? ` (${bindingLabel})` : '');
             return {
