+++ webapp/packages/plugin-object-viewer-nav-tree-link/tsconfig.json
@@ -20,6 +20,9 @@
       "path": "../core-localization"
     },
     {
+      "path": "../core-utils"
+    },
+    {
       "path": "../core-view"
     },
     {
